First prompt: I want to build a static webapp that uses the HTML5 filesystem API to allow the user to select a local directory containing images. The user should be presented with a single image at a time with the option to classify the number showing on the image, which could be from a 4-sided, 6-sided, 8-sided, 10-sided, 12-sided, or 20-sided die. There should also be an option to classify the image as invalid.
When the user makes the selection, it should move the image into a new subdirectory named according to the number selected (or invalid for invalid). It should then show the next uncategorized image. It should also show an undo button, which will move the last categorized image back to its original location and make it the active image to review.